Move Page List tests to dedicated non Navigation spec
